Update UpdateNotificationLogic logic and add tests
