+++ src/it/java/teammates/it/sqllogic/core/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Contains test cases for {@link teammates.storage.sqlapi} package.
  */
-package teammates.it.storage.sqlapi;
+package teammates.it.sqllogic.core;
