+++ src/main/java/teammates/storage/sqlapi/NotificationsDb.java
@@ -1,12 +1,8 @@
 package teammates.storage.sqlapi;
 
-import java.time.Instant;
 import java.util.UUID;
 
-import teammates.common.datatransfer.NotificationStyle;
-import teammates.common.datatransfer.NotificationTargetUser;
 import teammates.common.exception.EntityAlreadyExistsException;
-import teammates.common.exception.EntityDoesNotExistException;
 import teammates.common.exception.InvalidParametersException;
 import teammates.common.util.HibernateUtil;
 import teammates.storage.sqlentity.Notification;
@@ -50,35 +46,6 @@ public final class NotificationsDb extends EntitiesDb<Notification> {
         assert notificationId != null;
 
         return HibernateUtil.get(Notification.class, notificationId);
-    }
-
-    /**
-     * Updates a notification with {@link Notification}.
-     */
-    public Notification updateNotification(UUID notificationId, Instant startTime, Instant endTime,
-                                           NotificationStyle style, NotificationTargetUser targetUser, String title,
-                                           String message)
-            throws InvalidParametersException, EntityDoesNotExistException {
-        assert notificationId != null;
-
-        Notification notification = getNotification(notificationId);
-
-        if (notification == null) {
-            throw new EntityDoesNotExistException(ERROR_UPDATE_NON_EXISTENT);
-        }
-
-        notification.setStartTime(startTime);
-        notification.setEndTime(endTime);
-        notification.setStyle(style);
-        notification.setTargetUser(targetUser);
-        notification.setTitle(title);
-        notification.setMessage(message);
-
-        if (!notification.isValid()) {
-            throw new InvalidParametersException(notification.getInvalidityInfo());
-        }
-
-        return notification;
     }
 
     /**
