+++ src/test/java/teammates/sqllogic/core/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains test cases for {@link teammates.sqllogic.core} package.
+ */
+package teammates.sqllogic.core;
